+++ packages/blocks/core/js/third-party/blocksy-share-box/test/share-box-block.blocksy.e2e.cy.js
@@ -1,0 +1,935 @@
+/**
+ * Blockera dependencies
+ */
+import {
+	savePage,
+	redirectToFrontPage,
+	appendBlocks,
+	getSelectedBlock,
+	getWPDataObject,
+	createPost,
+	setInnerBlock,
+	setBlockState,
+	setParentBlock,
+} from '@blockera/dev-cypress/js/helpers';
+
+describe('Blocksy → Share Box Block → WP Compatibility', () => {
+	beforeEach(() => {
+		createPost();
+	});
+
+	describe('BG Color Inner Block', () => {
+		describe('Data Compatibility', () => {
+			it('Simple Value', () => {
+				appendBlocks(
+					`<!-- wp:blocksy/widgets-wrapper {"heading":"Share Box","block":"blocksy/share-box"} -->
+<!-- wp:heading {"level":3,"className":"","fontSize":"medium"} -->
+<h3 class="wp-block-heading has-medium-font-size">Share Box</h3>
+<!-- /wp:heading -->
+
+<!-- wp:blocksy/share-box {"share_type":"rounded","share_icons_fill":"solid","customBackgroundColor":"#7790ff","lock":{"remove":true}} -->
+<div>Blocksy: Share Box</div>
+<!-- /wp:blocksy/share-box -->
+<!-- /wp:blocksy/widgets-wrapper -->`
+				);
+
+				// Select target block
+				cy.getBlock('blocksy/share-box').first().click();
+
+				//
+				// Test 1: WP data to Blockera
+				//
+
+				// WP data should come to Blockera
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'backgroundColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customBackgroundColor')
+					);
+
+					expect('#7790ff').to.be.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBackgroundColor
+					);
+				});
+
+				//
+				// Test 2: Blockera value to WP data
+				//
+				setInnerBlock('elements/icons');
+
+				cy.setColorControlValue('BG Color', '666666');
+
+				//
+				// Check
+				//
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'backgroundColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customBackgroundColor')
+					);
+
+					expect('#666666').to.be.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBackgroundColor
+					);
+				});
+
+				//
+				// Test 3: Clear Blockera value and check WP data
+				//
+
+				cy.resetBlockeraAttribute('Background', 'BG Color', 'reset');
+
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'backgroundColor')
+					);
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customBackgroundColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBackgroundColor
+					);
+				});
+			});
+
+			it('Variable Value', () => {
+				appendBlocks(
+					`<!-- wp:blocksy/widgets-wrapper {"heading":"Share Box","block":"blocksy/share-box"} -->
+<!-- wp:heading {"level":3,"className":"","fontSize":"medium"} -->
+<h3 class="wp-block-heading has-medium-font-size">Share Box</h3>
+<!-- /wp:heading -->
+
+<!-- wp:blocksy/share-box {"share_type":"rounded","share_icons_fill":"solid","backgroundColor":"palette-color-1","lock":{"remove":true}} -->
+<div>Blocksy: Share Box</div>
+<!-- /wp:blocksy/share-box -->
+<!-- /wp:blocksy/widgets-wrapper -->`
+				);
+
+				// Select target block
+				cy.getBlock('blocksy/share-box').first().click();
+
+				//
+				// Test 1: WP data to Blockera
+				//
+
+				// WP data should come to Blockera
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'backgroundColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customBackgroundColor')
+					);
+
+					expect({
+						settings: {
+							name: 'Palette Color 1',
+							id: 'palette-color-1',
+							value: 'var(--theme-palette-color-1, #2872fa)',
+							reference: {
+								type: 'theme',
+								theme: 'Blocksy',
+							},
+							type: 'color',
+							var: '--wp--preset--color--palette-color-1',
+						},
+						name: 'Palette Color 1',
+						isValueAddon: true,
+						valueType: 'variable',
+					}).to.be.deep.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBackgroundColor
+					);
+				});
+
+				//
+				// Test 2: Blockera value to WP data
+				//
+				setInnerBlock('elements/icons');
+
+				cy.getParentContainer('BG Color').within(() => {
+					cy.clickValueAddonButton();
+				});
+
+				// change variable
+				cy.selectValueAddonItem('palette-color-2');
+
+				//
+				// Check
+				//
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'backgroundColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customBackgroundColor')
+					);
+
+					expect({
+						settings: {
+							name: 'Palette Color 2',
+							id: 'palette-color-2',
+							value: 'var(--theme-palette-color-2, #1559ed)',
+							reference: {
+								type: 'theme',
+								theme: 'Blocksy',
+							},
+							type: 'color',
+							var: '--wp--preset--color--palette-color-2',
+						},
+						name: 'Palette Color 2',
+						isValueAddon: true,
+						valueType: 'variable',
+					}).to.be.deep.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBackgroundColor
+					);
+				});
+
+				//
+				// Test 3: Clear Blockera value and check WP data
+				//
+
+				cy.getParentContainer('BG Color').within(() => {
+					cy.removeValueAddon();
+				});
+
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'backgroundColor')
+					);
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customBackgroundColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBackgroundColor
+					);
+				});
+			});
+		});
+	});
+
+	describe('BG Hover Color Inner Block', () => {
+		describe('Data Compatibility', () => {
+			it('Simple Value', () => {
+				appendBlocks(
+					`<!-- wp:blocksy/widgets-wrapper {"heading":"Share Box","block":"blocksy/share-box"} -->
+<!-- wp:heading {"level":3,"className":"","fontSize":"medium"} -->
+<h3 class="wp-block-heading has-medium-font-size">Share Box</h3>
+<!-- /wp:heading -->
+
+<!-- wp:blocksy/share-box {"share_type":"rounded","share_icons_fill":"solid","customBackgroundHoverColor":"#ff0000","lock":{"remove":true}} -->
+<div>Blocksy: Share Box</div>
+<!-- /wp:blocksy/share-box -->
+<!-- /wp:blocksy/widgets-wrapper -->`
+				);
+
+				// Select target block
+				cy.getBlock('blocksy/share-box').first().click();
+
+				//
+				// Test 1: WP data to Blockera
+				//
+
+				// WP data should come to Blockera
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'backgroundHoverColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customBackgroundHoverColor')
+					);
+
+					expect('#ff0000').to.be.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBlockStates?.hover?.breakpoints
+							.desktop?.attributes?.blockeraBackgroundColor
+					);
+				});
+
+				//
+				// Test 2: Blockera value to WP data
+				//
+				setInnerBlock('elements/icons');
+				setBlockState('Hover');
+
+				cy.setColorControlValue('BG Color', '666666');
+
+				//
+				// Check
+				//
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'backgroundHoverColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customBackgroundHoverColor')
+					);
+
+					expect('#666666').to.be.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBlockStates?.hover?.breakpoints
+							.desktop?.attributes?.blockeraBackgroundColor
+					);
+				});
+
+				//
+				// Test 3: Clear Blockera value and check WP data
+				//
+
+				cy.resetBlockeraAttribute('Background', 'BG Color', 'reset');
+
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'backgroundHoverColor')
+					);
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customBackgroundHoverColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBlockStates?.hover?.breakpoints
+							.desktop?.attributes?.blockeraBackgroundColor
+					);
+				});
+			});
+
+			it('Variable Value', () => {
+				appendBlocks(
+					`<!-- wp:blocksy/widgets-wrapper {"heading":"Share Box","block":"blocksy/share-box"} -->
+<!-- wp:heading {"level":3,"className":"","fontSize":"medium"} -->
+<h3 class="wp-block-heading has-medium-font-size">Share Box</h3>
+<!-- /wp:heading -->
+
+<!-- wp:blocksy/share-box {"share_type":"rounded","share_icons_fill":"solid","backgroundHoverColor":"palette-color-1","lock":{"remove":true}} -->
+<div>Blocksy: Share Box</div>
+<!-- /wp:blocksy/share-box -->
+<!-- /wp:blocksy/widgets-wrapper -->`
+				);
+
+				// Select target block
+				cy.getBlock('blocksy/share-box').first().click();
+
+				//
+				// Test 1: WP data to Blockera
+				//
+
+				// WP data should come to Blockera
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'backgroundHoverColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customBackgroundHoverColor')
+					);
+
+					expect({
+						settings: {
+							name: 'Palette Color 1',
+							id: 'palette-color-1',
+							value: 'var(--theme-palette-color-1, #2872fa)',
+							reference: {
+								type: 'theme',
+								theme: 'Blocksy',
+							},
+							type: 'color',
+							var: '--wp--preset--color--palette-color-1',
+						},
+						name: 'Palette Color 1',
+						isValueAddon: true,
+						valueType: 'variable',
+					}).to.be.deep.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBlockStates?.hover?.breakpoints
+							.desktop?.attributes?.blockeraBackgroundColor
+					);
+				});
+
+				//
+				// Test 2: Blockera value to WP data
+				//
+				setInnerBlock('elements/icons');
+				setBlockState('Hover');
+
+				cy.getParentContainer('BG Color').within(() => {
+					cy.clickValueAddonButton();
+				});
+
+				// change variable
+				cy.selectValueAddonItem('palette-color-2');
+
+				//
+				// Check
+				//
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'backgroundHoverColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customBackgroundHoverColor')
+					);
+
+					expect({
+						settings: {
+							name: 'Palette Color 2',
+							id: 'palette-color-2',
+							value: 'var(--theme-palette-color-2, #1559ed)',
+							reference: {
+								type: 'theme',
+								theme: 'Blocksy',
+							},
+							type: 'color',
+							var: '--wp--preset--color--palette-color-2',
+						},
+						name: 'Palette Color 2',
+						isValueAddon: true,
+						valueType: 'variable',
+					}).to.be.deep.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBlockStates?.hover?.breakpoints
+							.desktop?.attributes?.blockeraBackgroundColor
+					);
+				});
+
+				//
+				// Test 3: Clear Blockera value and check WP data
+				//
+
+				cy.getParentContainer('BG Color').within(() => {
+					cy.removeValueAddon();
+				});
+
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'backgroundHoverColor')
+					);
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customBackgroundHoverColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBlockStates?.hover?.breakpoints
+							.desktop?.attributes?.blockeraBackgroundColor
+					);
+				});
+			});
+		});
+	});
+
+	describe('Icons Color Inner Block', () => {
+		describe('Data Compatibility', () => {
+			it('Simple Value', () => {
+				appendBlocks(
+					`<!-- wp:blocksy/widgets-wrapper {"heading":"Share Box","block":"blocksy/share-box"} -->
+<!-- wp:heading {"level":3,"className":"","fontSize":"medium"} -->
+<h3 class="wp-block-heading has-medium-font-size">Share Box</h3>
+<!-- /wp:heading -->
+
+<!-- wp:blocksy/share-box {"customInitialColor":"#ff9e9e","lock":{"remove":true}} -->
+<div>Blocksy: Share Box</div>
+<!-- /wp:blocksy/share-box -->
+<!-- /wp:blocksy/widgets-wrapper -->`
+				);
+
+				// Select target block
+				cy.getBlock('blocksy/share-box').first().click();
+
+				//
+				// Test 1: WP data to Blockera
+				//
+
+				// WP data should come to Blockera
+				getWPDataObject().then((data) => {
+					expect('').to.be.equal(
+						getSelectedBlock(data, 'initialColor')
+					);
+
+					expect('#ff9e9e').to.be.equal(
+						getSelectedBlock(data, 'customInitialColor')
+					);
+
+					expect('#ff9e9e').to.be.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraFontColor
+					);
+				});
+
+				//
+				// Test 2: Blockera value to WP data
+				//
+				setInnerBlock('elements/icons');
+
+				cy.setColorControlValue('Text Color', '666666');
+
+				//
+				// Check
+				//
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'initialColor')
+					);
+
+					expect('#666666').to.be.equal(
+						getSelectedBlock(data, 'customInitialColor')
+					);
+
+					expect('#666666').to.be.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraFontColor
+					);
+				});
+
+				//
+				// Test 3: Clear Blockera value and check WP data
+				//
+
+				cy.resetBlockeraAttribute('Typography', 'Text Color', 'reset');
+
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'initialColor')
+					);
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customInitialColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraFontColor
+					);
+				});
+			});
+
+			it('Variable Value', () => {
+				appendBlocks(
+					`<!-- wp:blocksy/widgets-wrapper {"heading":"Share Box","block":"blocksy/share-box"} -->
+<!-- wp:heading {"level":3,"className":"","fontSize":"medium"} -->
+<h3 class="wp-block-heading has-medium-font-size">Share Box</h3>
+<!-- /wp:heading -->
+
+<!-- wp:blocksy/share-box {"initialColor":"palette-color-1","lock":{"remove":true}} -->
+<div>Blocksy: Share Box</div>
+<!-- /wp:blocksy/share-box -->
+<!-- /wp:blocksy/widgets-wrapper -->`
+				);
+
+				// Select target block
+				cy.getBlock('blocksy/share-box').first().click();
+
+				//
+				// Test 1: WP data to Blockera
+				//
+
+				// WP data should come to Blockera
+				getWPDataObject().then((data) => {
+					expect('palette-color-1').to.be.equal(
+						getSelectedBlock(data, 'initialColor')
+					);
+
+					expect('').to.be.equal(
+						getSelectedBlock(data, 'customInitialColor')
+					);
+
+					expect({
+						settings: {
+							name: 'Palette Color 1',
+							id: 'palette-color-1',
+							value: 'var(--theme-palette-color-1, #2872fa)',
+							reference: {
+								type: 'theme',
+								theme: 'Blocksy',
+							},
+							type: 'color',
+							var: '--wp--preset--color--palette-color-1',
+						},
+						name: 'Palette Color 1',
+						isValueAddon: true,
+						valueType: 'variable',
+					}).to.be.deep.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraFontColor
+					);
+				});
+
+				//
+				// Test 2: Blockera value to WP data
+				//
+				setInnerBlock('elements/icons');
+
+				cy.getParentContainer('Text Color').within(() => {
+					cy.clickValueAddonButton();
+				});
+
+				// change variable
+				cy.selectValueAddonItem('palette-color-2');
+
+				//
+				// Check
+				//
+				getWPDataObject().then((data) => {
+					expect('palette-color-2').to.be.equal(
+						getSelectedBlock(data, 'initialColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customInitialColor')
+					);
+
+					expect({
+						settings: {
+							name: 'Palette Color 2',
+							id: 'palette-color-2',
+							value: 'var(--theme-palette-color-2, #1559ed)',
+							reference: {
+								type: 'theme',
+								theme: 'Blocksy',
+							},
+							type: 'color',
+							var: '--wp--preset--color--palette-color-2',
+						},
+						name: 'Palette Color 2',
+						isValueAddon: true,
+						valueType: 'variable',
+					}).to.be.deep.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraFontColor
+					);
+				});
+
+				//
+				// Test 3: Clear Blockera value and check WP data
+				//
+
+				cy.getParentContainer('Text Color').within(() => {
+					cy.removeValueAddon();
+				});
+
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'initialColor')
+					);
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customInitialColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraFontColor
+					);
+				});
+			});
+		});
+	});
+
+	describe('Icons Hover Color Inner Block', () => {
+		describe('Data Compatibility', () => {
+			it('Simple Value', () => {
+				appendBlocks(
+					`<!-- wp:blocksy/widgets-wrapper {"heading":"Share Box","block":"blocksy/share-box"} -->
+<!-- wp:heading {"level":3,"className":"","fontSize":"medium"} -->
+<h3 class="wp-block-heading has-medium-font-size">Share Box</h3>
+<!-- /wp:heading -->
+
+<!-- wp:blocksy/share-box {"customHoverColor":"#ff1c1c","lock":{"remove":true}} -->
+<div>Blocksy: Share Box</div>
+<!-- /wp:blocksy/share-box -->
+<!-- /wp:blocksy/widgets-wrapper -->`
+				);
+
+				// Select target block
+				cy.getBlock('blocksy/share-box').first().click();
+
+				//
+				// Test 1: WP data to Blockera
+				//
+
+				// WP data should come to Blockera
+				getWPDataObject().then((data) => {
+					expect('').to.be.equal(
+						getSelectedBlock(data, 'hoverColor')
+					);
+
+					expect('#ff1c1c').to.be.equal(
+						getSelectedBlock(data, 'customHoverColor')
+					);
+
+					expect('#ff1c1c').to.be.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBlockStates?.hover?.breakpoints
+							.desktop?.attributes?.blockeraFontColor
+					);
+				});
+
+				//
+				// Test 2: Blockera value to WP data
+				//
+				setInnerBlock('elements/icons');
+				setBlockState('Hover');
+
+				cy.setColorControlValue('Text Color', '666666');
+
+				//
+				// Check
+				//
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'hoverColor')
+					);
+
+					expect('#666666').to.be.equal(
+						getSelectedBlock(data, 'customHoverColor')
+					);
+
+					expect('#666666').to.be.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBlockStates?.hover?.breakpoints
+							.desktop?.attributes?.blockeraFontColor
+					);
+				});
+
+				//
+				// Test 3: Clear Blockera value and check WP data
+				//
+
+				cy.resetBlockeraAttribute('Typography', 'Text Color', 'reset');
+
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'hoverColor')
+					);
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customHoverColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBlockStates?.hover?.breakpoints
+							.desktop?.attributes?.blockeraFontColor
+					);
+				});
+			});
+
+			it('Variable Value', () => {
+				appendBlocks(
+					`<!-- wp:blocksy/widgets-wrapper {"heading":"Share Box","block":"blocksy/share-box"} -->
+<!-- wp:heading {"level":3,"className":"","fontSize":"medium"} -->
+<h3 class="wp-block-heading has-medium-font-size">Share Box</h3>
+<!-- /wp:heading -->
+
+<!-- wp:blocksy/share-box {"hoverColor":"palette-color-1","lock":{"remove":true}} -->
+<div>Blocksy: Share Box</div>
+<!-- /wp:blocksy/share-box -->
+<!-- /wp:blocksy/widgets-wrapper -->`
+				);
+
+				// Select target block
+				cy.getBlock('blocksy/share-box').first().click();
+
+				//
+				// Test 1: WP data to Blockera
+				//
+
+				// WP data should come to Blockera
+				getWPDataObject().then((data) => {
+					expect('palette-color-1').to.be.equal(
+						getSelectedBlock(data, 'hoverColor')
+					);
+
+					expect('').to.be.equal(
+						getSelectedBlock(data, 'customHoverColor')
+					);
+
+					expect({
+						settings: {
+							name: 'Palette Color 1',
+							id: 'palette-color-1',
+							value: 'var(--theme-palette-color-1, #2872fa)',
+							reference: {
+								type: 'theme',
+								theme: 'Blocksy',
+							},
+							type: 'color',
+							var: '--wp--preset--color--palette-color-1',
+						},
+						name: 'Palette Color 1',
+						isValueAddon: true,
+						valueType: 'variable',
+					}).to.be.deep.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBlockStates?.hover?.breakpoints
+							.desktop?.attributes?.blockeraFontColor
+					);
+				});
+
+				//
+				// Test 2: Blockera value to WP data
+				//
+				setInnerBlock('elements/icons');
+				setBlockState('Hover');
+
+				cy.getParentContainer('Text Color').within(() => {
+					cy.clickValueAddonButton();
+				});
+
+				// change variable
+				cy.selectValueAddonItem('palette-color-2');
+
+				//
+				// Check
+				//
+				getWPDataObject().then((data) => {
+					expect('palette-color-2').to.be.equal(
+						getSelectedBlock(data, 'hoverColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customHoverColor')
+					);
+
+					expect({
+						settings: {
+							name: 'Palette Color 2',
+							id: 'palette-color-2',
+							value: 'var(--theme-palette-color-2, #1559ed)',
+							reference: {
+								type: 'theme',
+								theme: 'Blocksy',
+							},
+							type: 'color',
+							var: '--wp--preset--color--palette-color-2',
+						},
+						name: 'Palette Color 2',
+						isValueAddon: true,
+						valueType: 'variable',
+					}).to.be.deep.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBlockStates?.hover?.breakpoints
+							.desktop?.attributes?.blockeraFontColor
+					);
+				});
+
+				//
+				// Test 3: Clear Blockera value and check WP data
+				//
+
+				cy.getParentContainer('Text Color').within(() => {
+					cy.removeValueAddon();
+				});
+
+				getWPDataObject().then((data) => {
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'hoverColor')
+					);
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'customHoverColor')
+					);
+
+					expect(undefined).to.be.equal(
+						getSelectedBlock(data, 'blockeraInnerBlocks')[
+							'elements/icons'
+						]?.attributes?.blockeraBlockStates?.hover?.breakpoints
+							.desktop?.attributes?.blockeraFontColor
+					);
+				});
+			});
+		});
+	});
+
+	it('Inner blocks existence + CSS selectors in editor and front-end', () => {
+		appendBlocks(`<!-- wp:blocksy/widgets-wrapper {"heading":"Share Box","block":"blocksy/share-box"} -->
+<!-- wp:heading {"level":3,"className":"","fontSize":"medium"} -->
+<h3 class="wp-block-heading has-medium-font-size">Share Box</h3>
+<!-- /wp:heading -->
+
+<!-- wp:blocksy/share-box {"blockeraPropsId":"b8ef75a7-b1b7-49b0-b60e-a61976db9e8e","blockeraCompatId":"18121630115","blockeraBackgroundColor":{"value":"#ffdddd"},"blockeraBorder":{"value":{"type":"all","all":{"width":"","color":"rgba(218, 222, 228, 0.5)","style":"solid"}}},"blockeraInnerBlocks":{"value":{"elements/icons":{"attributes":{"blockeraBackgroundColor":"#ff0000"}}}},"share_networks":[{"id":"facebook","enabled":true,"__id":"JZgkASt-GgZuyrK_JXVm0"},{"id":"twitter","enabled":true,"__id":"PGAd9qyLle5-FtosY0Is8"},{"id":"pinterest","enabled":true,"__id":"Jq6k3oHXcsG_ac4gDc2ax"},{"id":"linkedin","enabled":true,"__id":"LybrPWpSWaVCtFn5boSDW"}],"share_item_tooltip":"yes","link_nofollow":"yes","share_icons_size":"31","lock":{"remove":true},"className":"blockera-block blockera-block\u002d\u002d6kc8u4"} -->
+<div>Blocksy: Share Box</div>
+<!-- /wp:blocksy/share-box -->
+<!-- /wp:blocksy/widgets-wrapper -->`);
+
+		cy.getBlock('blocksy/share-box').first().click();
+
+		cy.get('.blockera-extension-block-card').should('be.visible');
+
+		//
+		//  1. Assert inner blocks selectors in editor
+		//
+		cy.getBlock('blocksy/share-box').should(
+			'have.css',
+			'background-color',
+			'rgb(255, 221, 221)'
+		);
+
+		cy.getBlock('blocksy/share-box')
+			.first()
+			.within(() => {
+				cy.get('.ct-icon-container').should(
+					'have.css',
+					'background-color',
+					'rgb(255, 0, 0)'
+				);
+			});
+
+		//
+		// 2. Assert inner blocks selectors in front end
+		//
+		savePage();
+		redirectToFrontPage();
+
+		cy.get('.blockera-block').should(
+			'have.css',
+			'background-color',
+			'rgb(255, 221, 221)'
+		);
+
+		cy.get('.blockera-block').within(() => {
+			cy.get('.ct-icon-container').should(
+				'have.css',
+				'background-color',
+				'rgb(255, 0, 0)'
+			);
+		});
+	});
+});